refactor: Extract duplicated type-safe accessor utilities
- Create TypeSafeAccessorUtils object to eliminate code duplication
- Move normalizeTypeSafeProjectAccessor() and isTypeSafeProjectAccessor() to shared utility
- Update AdviceFinder and GroovyBuildScriptDependenciesRewriter to use utilities
- Maintain functionality while improving maintainability
